+++ src/CrossyToad/Physics/Distance.hs
@@ -1,4 +1,4 @@
 module CrossyToad.Physics.Distance where
 
 -- | How far something is in pixels
-type Distance = Float
+type Distance = Double
