+++ src/CrossyToad/Physics/Speed.hs
@@ -1,4 +1,4 @@
 module CrossyToad.Physics.Speed where
 
 -- | How fast an object moves per second in pixels
-type Speed = Float
+type Speed = Double
